configure logging for the full package
