+++ omf/scripts/geoh5_to_omf.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from omf.fileio import OMFWriter
 from omf.fileio.geoh5 import GeoH5Reader
 
-logger = logging.getLogger(__package__)
+_logger = logging.getLogger(__package__)
 
 
 def run():
@@ -17,14 +17,14 @@ def run():
         if not output_filepath.suffix:
             output_filepath = output_filepath.with_suffix(".omf")
     if output_filepath.exists():
-        logger.error(
+        _logger.error(
             f"Cowardly refuses to overwrite existing file '{output_filepath}'."
         )
         exit(1)
 
     reader = GeoH5Reader(geoh5_filepath)
     OMFWriter(reader(), str(output_filepath.absolute()))
-    logger.info(f"geoh5 file created: {output_filepath}")
+    _logger.info(f"OMF file created: {output_filepath}")
 
 
 if __name__ == "__main__":
