+++ omf/scripts/omf_to_geoh5.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from omf.fileio import OMFReader
 from omf.fileio.geoh5 import GeoH5Writer
 
-logger = logging.getLogger(__package__)
+_logger = logging.getLogger(__package__)
 
 
 def run():
@@ -18,14 +18,14 @@ def run():
         if not output_filepath.suffix:
             output_filepath = output_filepath.with_suffix(".geoh5")
     if output_filepath.exists():
-        logger.error(
+        _logger.error(
             f"Cowardly refuses to overwrite existing file '{output_filepath}'."
         )
         exit(1)
 
     reader = OMFReader(str(omf_filepath.absolute()))
     GeoH5Writer(reader.get_project(), output_filepath)
-    logger.info(f"OMF file created: {output_filepath}")
+    _logger.info(f"geoh5 file created: {output_filepath}")
 
 
 if __name__ == "__main__":
